python_wikipedia_mobile/model/app.py
from selene import be
from selene.support.shared import browser
from appium.webdriver.common.appiumby import AppiumBy


def given_opened():
    if browser.element((AppiumBy.ID, 'fragment_onboarding_skip_button')).matching(be.visible):
        browser.element((AppiumBy.ID, 'fragment_onboarding_skip_button')).click()



